+++ changelog.md
@@ -3,6 +3,7 @@
 *<c-aaaaaa>Friendly reminder that</c> <cl>some</c> <cj>features</c> <c-aaaaaa>of this mod (including features from this update and features related to bugfixes in this update) require [Better Touch Prio](mod:alk.better-touch-prio) to be installed. These decisions are still here to stay.</c>*
 - <cg>Added</c> playback increment/decrement controls.
   - You can toggle this option off in the mod settings.
+  - You can also adjust how many milliseconds to increment/decrement by in the mod settings.
   - [Better Touch Prio](mod:alk.better-touch-prio) is <cl>requ</c><cj>ired</c> for this feature.
 - <co>Fixed</c> an issue where menu loop positions wouldn't be restored when leaving a level whose most recently played song was also the most recently selected menu loop.
 ## v1.11.3, v1.11.4, and v1.11.5